Replace camera system
This commit replaces the previous camera system with a system that keeps
track of which room each robber corresponds to. Whenever a robber moves to
a new room, that robber will be "attached" to the new room. If that robber
also happens to be selected, the camera for the room the robber is attached to
will be enabled. Furthermore, if a new robber is selected, the room that
the new robber is in will have its camera enabled instead.